+++ libs/core/shellbar/user-menu/shellbar-user-menu-button.directive.ts
@@ -1,0 +1,29 @@
+import { Directive, inject, Input, isDevMode, OnInit } from '@angular/core';
+import { defaultButtonType, FD_BUTTON_COMPONENT } from '@fundamental-ngx/core/button';
+
+@Directive({
+    selector: '[fd-button][fdShellbarUserMenuButton]',
+    host: {
+        '[class.fd-shellbar__button]': 'true',
+        '[class.fd-shellbar__button--user-menu]': 'true'
+    },
+    standalone: true
+})
+export class ShellbarUserMenuButtonDirective implements OnInit {
+    /** The type of the button. In case of the ShellbarUserMenuButtonDirective, the type is always 'transparent'. */
+    @Input()
+    fdType = 'transparent' as const;
+
+    /** @hidden */
+    private _hostButton = inject(FD_BUTTON_COMPONENT, { host: true });
+
+    /** @hidden */
+    ngOnInit(): void {
+        if (this._hostButton.fdType !== 'transparent' && this._hostButton.fdType !== defaultButtonType && isDevMode()) {
+            console.warn(
+                'ShellbarUserMenuButtonDirective should be used with transparent button type. Overriding button type to transparent.'
+            );
+        }
+        this._hostButton.fdType = 'transparent';
+    }
+}